+++ lab_07_05/hasho.h
@@ -20,7 +20,7 @@ typedef ho_t *ho_pt;
 
 hoe_pt hoe_init(char *k , char *v);
 ho_pt ho_init(int n);
-int ho_hash(ho_pt h, char *k);
+uint64_t ho_hash(ho_pt h, char *k);
 int ho_add(ho_pt h, char *k, char *v);
 char *ho_get(ho_pt h, char *k);
 void ho_del(ho_pt h, char *k);
